% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenEval.R
\name{ConfIntMeanBoot}
\alias{ConfIntMeanBoot}
\title{Bootstrap confidence interval}
\usage{
ConfIntMeanBoot(x, levels = NULL, conf = 0.95, n.rep = 10^4)
}
\arguments{
\item{x}{A vector containing values to calculate confidence interval.}

\item{levels}{A vector containing the levels (factor) from the data, to calculate separate confidence intervals.}

\item{n.rep}{The number of bootstrap replicates.}

\item{p}{The confidence level of the interval.}
}
\value{
A data frame with the confidence interals.
}
\description{
This function calculates the bootstrap confidence interval for the mean of a set of data.
}
