% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenEval.R
\name{ConfIntMeanNorm}
\alias{ConfIntMeanNorm}
\title{Confidence interval for normal distributed data}
\usage{
ConfIntMeanNorm(x, levels = NULL, p = 0.95)
}
\arguments{
\item{x}{A vector containing values to calculate confidence interval.}

\item{levels}{A vector containing the levels (factor) from the data, to calculate separate confidence intervals.}

\item{p}{The confidence level of the interval.}
}
\value{
A data frame with the confidence interals.
}
\description{
This function calculates the confidence interval for the mean of a set of normal distributed data.
}
