% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenEval.R
\name{LimPredAcc}
\alias{LimPredAcc}
\title{Calculates the expected limit of prediction accuracy}
\usage{
LimPredAcc(a, b, n, h2, p = 0.95)
}
\arguments{
\item{a}{The intercept obtained with \code{popCor}.}

\item{b}{The slope obtained with \code{popCor}.}

\item{n}{The size of the test population (the population for which predictions are to be made).}

\item{h2}{The heritability of the trait.}

\item{p}{The confidence level of the interval for the expected limit of prediction accuracy.}
}
\value{
A data frame with the mean expected limit of prediction accuracy, as well as the boundaries of its confidence interval.
}
\description{
This function calculates the expected limit of prediction accuracy, using the relationship between two populations based on the singular-value decomposition of their genotype matrices. This limit is obtained as in Cuyabano and Gondro, 2020 (in prep), and the two populations must have the SNP-genotypes on the same set of SNPs.
}
\examples{
# simulate genotypes
M <- simGeno(n=500,AlleleFreq=rep(0.5,100))
# separate two populations of individuals
# note that the populations do not need to have the same size
M1 <- M[1:300,]
M2 <- M[301:500,]
tmp <- popCor(M1,M2)
LimPredAcc(a=tmp$a,b=tmp$b,n=200,h2=0.5)
LimPredAcc(a=tmp$a,b=tmp$b,n=200,h2=seq(0.1,0.9,0.1))
# compare the expected limits of prediction accuracy to the classic sqrt(h2)
par(mar=c(4.5,4.5,1,1))
curve(sqrt(x),xlim=c(0,1),ylim=c(0,1),xlab=expression(h^2),ylab="Limit of prediction accuracy")
aux <- LimPredAcc(a=tmp$a,b=tmp$b,n=200,h2=seq(0.01,0.99,0.01))
lines(seq(0.01,0.99,0.01),aux$mean,lty=2)
lines(seq(0.01,0.99,0.01),aux$CI_lower,lty=3,col=2)
lines(seq(0.01,0.99,0.01),aux$CI_upper,lty=3,col=2)
legend("topleft",lty=c(1,2,3),col=c(1,1,2),legend=c("sqrt(h2)","R_expected","R_CI"))
}
