% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenEval.R
\name{MatrixInv}
\alias{MatrixInv}
\title{Inverse of a matrix}
\usage{
MatrixInv(A, use.eigen = FALSE, U = NULL, L = NULL)
}
\arguments{
\item{A}{A matrix.}

\item{use.eigen}{A logical indicating if the eigen-decomposition should be used to obtain the inverse. Valid only for symmetric matrices.}

\item{U}{A matrix of the eigenvectors (optional).}

\item{L}{A vector with the eigenvalues (optional).}
}
\value{
The inverse of the matrix.
}
\description{
This function inverts any matrix, using the best method for that specific matrix.
}
