% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenEval.R
\name{mkGRM}
\alias{mkGRM}
\title{Genomic relationship matrix}
\usage{
mkGRM(M, method = "VanRaden")
}
\arguments{
\item{M}{A SNP-genotype matrix (unphased).}

\item{method}{The method to use for standardizing the SNP-genotypes. Deafult is \code{method="VanRaden"}, and alternatively can be set \code{method="Individual"}, which will standardize each SNP individually.}
}
\value{
The genomic relationship matrix.
}
\description{
This function builds the genomic relationship matrix based on SNP-genotypes.
}
\examples{
M <- simGeno(5,rep(0.5,1000))
M <- M[,apply(M,2,var) != 0]
mkGRM(M)
mkGRM(M,method="Individual")
}
