% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenEval.R
\name{mkPED}
\alias{mkPED}
\title{Pedigree relationship matrix}
\usage{
mkPED(pop.info)
}
\arguments{
\item{pop.info}{The population info, in the same format generated by \code{simPopInfo}.}
}
\value{
The pedigree relationship matrix.
}
\description{
This function builds the pedigree relationship matrix based on a population info in the same format as the generated by function \code{simPopInfo}. If you are generating the population info using \code{simPopInfo}, the pedigree relationship matrix can be created directly from \code{simPopInfo} by setting the parameter \code{pedigree=TRUE}.
}
\examples{
mkPED(simPopInfo(10,0))
}
