% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenEval.R
\name{plotMP}
\alias{plotMP}
\title{Plot eigen-values comparatively with The Marchenko-Pastur distribution}
\usage{
plotMP(L, a, s2, f = "cumulative")
}
\arguments{
\item{L}{A vectors of eigen-values.}

\item{a}{Rate parameter (n/m = #observations/#variables).}

\item{s2}{Variance parameter.}

\item{f}{The type of function to be plotted, \code{"density"} or \code{"cumulative"} (default).}
}
\description{
Plot eigen-values comparatively with The Marchenko-Pastur distribution
}
