% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenEval.R
\name{pMP}
\alias{pMP}
\title{The Marchenko-Pastur cumulative distribution}
\usage{
pMP(x, a, s2)
}
\arguments{
\item{x}{A vectors of quantiles.}

\item{a}{Rate parameter (n/m = #observations/#variables).}

\item{s2}{Variance parameter.}
}
\value{
The cumulative probability distribution.
}
\description{
Cumulative distribution function of the Marchenko-Pastur distribution with rate parameter equal to \code{a} and variance equal to \code{s2}.
}
