% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenEval.R
\name{popCor}
\alias{popCor}
\title{Calculates the relationship between two populations}
\usage{
popCor(M1, M2)
}
\arguments{
\item{M1}{The SNP-genotypes of population one.}

\item{M2}{The SNP-genotypes of population two.}
}
\value{
The parameters of the linear regression (intercept and slope) on the components built from the SVDs.
}
\description{
This function calculates the relationship between two populations based on the singular-value decomposition of their genotype matrices. The relationship is obtained as in Cuyabano and Gondro, 2020 (in prep.), and the two populations must have the SNP-genotypes on the same set of SNPs.
}
\examples{
# simulate genotypes
M <- simGeno(n=500,AlleleFreq=rep(0.5,100),LD=TRUE)
# separate two populations of individuals
# note that the populations do not need to have the same size
M1 <- M[1:300,]
M2 <- M[301:500,]
popCor(M1,M2)
}
