% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenEval.R
\name{REMLsolve}
\alias{REMLsolve}
\title{REML estimates of variance components}
\usage{
REMLsolve(y, X = NULL, Z = NULL, VarMat = NULL, R = NULL)
}
\arguments{
\item{y}{The response variable.}

\item{X}{A matrix of fixed effects.}

\item{Z}{The design matrix (or list of matrices) of the random effects. If all design matrices are an identity, this argument does not need to be provided.}

\item{VarMat}{The variance-covariance matrix (or list of matrices) of the random effects.}

\item{R}{A variance-covariance matrix for the residuals, if they are not independent and identically distributed (i.e. R != In)}
}
\value{
\describe{
\item{FixedEff}{The estimated fixed effects.}
\item{VarComp}{The estimated variance components for all random effects, and for the residual.}
}
}
\description{
This function estimates the variance components and the fixed effects in a mixed model, using the restricted maximum likelihood (REML).
}
\examples{
## SINGLE COMPONENT WITH FIXED EFFECTS ##
# simulate genotypes
M <- simGeno(n=500,AlleleFreq=runif(100,0.05,0.5))
# simulate phenotypes
pheno <- simPheno(M,0.6,Vy=10)
# add fixed effects
X <- cbind(rep(1,nrow(M)),runif(nrow(M)),rbinom(nrow(M),1,0.5))
b <- c(5,-1,2)
y <- pheno$y + as.numeric(X\%*\%b)
# REML
G <- mkGRM(M)
REMLsolve(y=y,X=X,VarMat=G)

## TWO VARIANCE COMPONENTS ##
# simulate genotypes
M <- simGeno(n=500,AlleleFreq=runif(100,0.05,0.5))
# simulate phenotypes
pheno <- simPheno(list(M[,1:70],M[,71:100]),c(0.4,0.2),Vy=10)
# add fixed effects
X <- cbind(rep(1,nrow(M)),runif(nrow(M)),rbinom(nrow(M),1,0.5))
b <- c(5,-1,2)
y <- pheno$y + as.numeric(X\%*\%b)
# REML
G <- list(mkGRM(M[,1:70]),mkGRM(M[,71:100]))
REMLsolve(y=y,X=X,VarMat=G)
REMLsolve(y=y,X=X,VarMat=list(diag(1,70),diag(1,30)),Z=list(M[,1:70],M[,71:100]))
tmp <- REMLsolve(y=y,X=X,VarMat=list(diag(1,70),diag(1,30)),Z=list(M[,1:70],M[,71:100]))$VarComp
c(sum(apply(M[,1:70],2,var)),sum(apply(M[,71:100],2,var)),1)*tmp
}
