% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenEval.R
\name{simGeno}
\alias{simGeno}
\title{Simulate genotypes}
\usage{
simGeno(n, AlleleFreq, LD = FALSE, R = NULL, phased = FALSE)
}
\arguments{
\item{n}{The number of individuals to simulate.}

\item{AlleleFreq}{A vector with the required allele frequencies for each SNP, with \code{length(AlleleFreq)} being the number of SNPs.}

\item{LD}{Logical indicating if SNP-genotypes should be simulated in linkage disequilibrium.}

\item{R}{The correlation matrix for the linkage disequilibrium between SNPs. If \code{R=NULL}, a default matrix with 0.95^(lag between SNPs) will be generated.}

\item{phased}{Lagical indicating if phased genotypes should be returned.}
}
\value{
The simulated SNPs, as a single matrix with allele counts (\code{phased=FALSE}), or as a list of matrices with t0,1 at each allele (\code{phased=TRUE}).
}
\description{
This function simulates numeric SNP-genotypes.
}
\examples{
simGeno(n=10,AlleleFreq=rep(0.5,5))
simGeno(n=10,AlleleFreq=rep(0.5,5),phased=TRUE)
simGeno(n=10,AlleleFreq=rep(0.5,5),LD=TRUE)
R_LD <- diag(1,5)
R_LD[abs(col(R_LD)-row(R_LD)) == 1] <- 0.95
R_LD[abs(col(R_LD)-row(R_LD)) == 2] <- 0.85
R_LD[abs(col(R_LD)-row(R_LD)) == 3] <- 0.75
R_LD[abs(col(R_LD)-row(R_LD)) == 4] <- 0.65
simGeno(n=10,AlleleFreq=rep(0.5,5),LD=TRUE,R=R_LD)
geno <- simGeno(n=1000,AlleleFreq=rep(0.5,5),LD=TRUE,R=R_LD,phased=TRUE)
str(geno)
cor(rbind(geno[[1]],geno[[2]]))
}
