% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenEval.R
\name{simPheno}
\alias{simPheno}
\title{Simulate phenotypes}
\usage{
simPheno(M, h2, Vy = 1, binary = FALSE, p = NULL, sim.error = 1e-04)
}
\arguments{
\item{M}{A matrix or a list of matrices of the (unphased) SNP-genotypes set to be the QTL.}

\item{h2}{The heritability.}

\item{Vy}{The phenotypic variance (for continuous traits). Default is \code{Vy=1}.}

\item{binary}{Logical indicating whether simulated phenotypes should be binary. Default is \code{binary=FALSE}.}

\item{p}{If \code{binary=TRUE}, the probability of observing 1's in the phenotypes, with default \code{p=0.5}.}

\item{sim.error}{An error threshold to be considered on the orthogonalization of genetic values and residuals.}
}
\value{
\describe{
\item{y}{The simulated phenotypes.}
\item{a}{The simulated QTL effects.}
\item{g}{The simulated genetic values.}
\item{e}{The simulated residuals.}
\item{h2}{The simulated heritability.}
}
}
\description{
This function simulates phenotypes (continuous or binary) based on the quantitative trait loci (QTL), only with the genetic values and residuals (\code{y=g+e}).
}
\examples{
## general examples ##
AF <- runif(500,0.05,1)
geno <- simGeno(n=1000,AlleleFreq=AF,LD=TRUE)
pheno <- simPheno(M=geno,h2=0.6,Vy=10)
str(pheno)
pheno <- simPheno(M=geno,h2=0.6,binary=TRUE)
str(pheno)
mean(pheno$y) # this is the realized probability of observing y = 1
pheno <- simPheno(M=geno,h2=0.6,binary=TRUE,p=0.2)
str(pheno)
mean(pheno$y) # this is the realized probability of observing y = 1

## example with multiple components with different heritabilities ##
AF <- runif(500,0.05,1)
geno <- simGeno(n=1000,AlleleFreq=AF,LD=TRUE)
pheno <- simPheno(M=list(geno[,1:350],geno[,351:500]),h2=c(0.2,0.4),Vy=10)
str(pheno)
pheno <- simPheno(M=list(geno[,1:350],geno[,351:500]),h2=c(0.2,0.4),Vy=10,binary=TRUE,p=0.3)
str(pheno)
mean(pheno$y) # this is the realized probability of observing y = 1
}
\seealso{
\code{\link{simGeno}}
}
